+++ src/meeting.ts
@@ -4,8 +4,6 @@ import {
 } from 'react-native-webrtc';
 import io from 'socket.io-client/dist/socket.io';
 import { MeetingUserModel } from "./Models/meeting-user-model";
-import { MeetingModel } from "./Models/meeting-model";
-import { ParticipantModel } from "./Models/participant-model";
 import { ConusmaException } from "./Exceptions/conusma-exception";
 import { ConusmaWorker } from "./conusma-worker";
 import InCallManager from 'react-native-incall-manager';
@@ -14,40 +12,28 @@ import { MediaServer } from "./media-server";
 import { Connection } from "./connection";
 
 
-
-export type MediaServerConnectionReadyObserver = () => void;
 export class Meeting {
-    public ownerUser: MeetingUserModel;
+    public activeUser: MeetingUserModel;
     public conusmaWorker: ConusmaWorker;
 
     public mediaServers: Array<MediaServer> = [];
     public connections: Array<Connection> = [];
 
-    private observers: MediaServerConnectionReadyObserver[] = [];
     private appService: AppService;
 
-    public isReceviedClose: boolean = false;
+    public isClosedRequestRecieved: boolean = false;
     
 
-    constructor(ownerUser: MeetingUserModel, appService: AppService) {
+    constructor(activeUser: MeetingUserModel, appService: AppService) {
         registerGlobals();
         this.appService = appService;
-        this.ownerUser = ownerUser;
-        this.conusmaWorker = new ConusmaWorker(this.appService, this.ownerUser);
-    }
-
-    public async createConnection(user: MeetingUserModel) {
-        const mediaServerModel: any = await this.getMediaServer(user.Id);
-        var mediaServer = await this.createMediaServer(mediaServerModel);
-        var connection = new Connection(user, mediaServer, this.appService);
-        this.mediaServers.push(mediaServer);
-        this.connections.push(connection);
-        return connection;
+        this.activeUser = activeUser;
+        this.conusmaWorker = new ConusmaWorker(this.appService, this.activeUser);
     }
 
     public startWorker() {
         try {
-            this.isReceviedClose = false;
+            this.isClosedRequestRecieved = false;
             this.conusmaWorker.start();
             this.conusmaWorker.meetingWorkerEvent.on('meetingUsers', () => {
                 console.log("Meeting users updated.");
@@ -67,10 +53,10 @@ export class Meeting {
     public async stopWorker(sendCloseRequest: boolean = false) {
         try {
             if (sendCloseRequest) {
-                var closeData = { 'MeetingUserId': this.ownerUser.Id };
+                var closeData = { 'MeetingUserId': this.activeUser.Id };
                 await this.appService.liveClose(closeData);
             }
-            this.isReceviedClose = true;
+            this.isClosedRequestRecieved = true;
             if (this.conusmaWorker != null) {
                 this.conusmaWorker.terminate();
             }
@@ -79,18 +65,6 @@ export class Meeting {
         }
     }
 
-    public attach(observer: MediaServerConnectionReadyObserver) {
-        this.observers.push(observer);
-    }
-
-    public detach(observerToRemove: MediaServerConnectionReadyObserver) {
-        this.observers = this.observers.filter(observer => observerToRemove !== observer);
-    }
-
-    private notify() {
-        this.observers.forEach(observer => observer());
-    }
-    
     private async getMediaServer(meetingUserId: string) {
         return await this.appService.getMediaServer(meetingUserId);
     }
@@ -103,13 +77,13 @@ export class Meeting {
             mediaServer.id = mediaServerModel.Id;
             mediaServer.socket = io.connect(mediaServerModel.ConnectionDnsAddress + ":" + mediaServerModel.Port);
             this.mediaServers.push(mediaServer);
-            var userInfoData = { 'MeetingUserId': this.ownerUser.Id, 'Token': this.appService.getJwtToken() };
+            var userInfoData = { 'MeetingUserId': this.activeUser.Id, 'Token': this.appService.getJwtToken() };
             let setUserInfo = await this.signal('UserInfo', userInfoData, mediaServer.socket);
 
         }
       
         mediaServer.socket.on('disconnect', async () => {
-            if (!this.isReceviedClose) {
+            if (!this.isClosedRequestRecieved) {
                 throw new ConusmaException("mediaserverconnection", "mediaserverconnection disconnect");
             }
         });
@@ -135,11 +109,8 @@ export class Meeting {
         else {
             console.error("no socket connection " + type);
         }
-
     }
     
-    
-
     public async enableAudioVideo() {
         try {
 
@@ -170,7 +141,7 @@ export class Meeting {
     }
     public async connectMeeting() {
         try {
-            await this.appService.connectMeeting(this.ownerUser);
+            await this.appService.connectMeeting(this.activeUser);
             console.log("User connected to the meeting.");
         } catch (error) {
             throw new ConusmaException("connectMeeting", "can not connect meeting , please check exception", error);
@@ -179,7 +150,7 @@ export class Meeting {
     }
     public async isApproved() {
         try {
-            return await this.appService.isApproved(this.ownerUser.Id);
+            return await this.appService.isApproved(this.activeUser.Id);
 
         } catch (error) {
             throw new ConusmaException("isApproved", "user is not approved, please check exception ", error);
@@ -197,8 +168,8 @@ export class Meeting {
     }
     public async getAllUsers() {
         try {
-            if (this.ownerUser != null) {
-                return <MeetingUserModel[]>await this.appService.getMeetingUserList({ 'MeetingUserId': this.ownerUser.Id });
+            if (this.activeUser != null) {
+                return <MeetingUserModel[]>await this.appService.getMeetingUserList({ 'MeetingUserId': this.activeUser.Id });
             } else {
                 return [];
             }
@@ -211,8 +182,8 @@ export class Meeting {
 
     public async getProducerUsers() {
         try {
-            if (this.ownerUser != null) {
-                var users = await this.appService.getMeetingUserList({ 'MeetingUserId': this.ownerUser.Id });
+            if (this.activeUser != null) {
+                var users = await this.appService.getMeetingUserList({ 'MeetingUserId': this.activeUser.Id });
                 var result: MeetingUserModel[] = [];
                 users.forEach((item: any) => {
                     if (item.Camera == true) {
@@ -227,5 +198,14 @@ export class Meeting {
             throw new ConusmaException("getProducerUsers", "Unable to fetch producer user list, please check detail exception");
         }
 
+    }
+    public setSpeaker(enable: boolean) {
+        try {
+            InCallManager.setSpeakerphoneOn(enable);
+            InCallManager.setForceSpeakerphoneOn(enable);
+
+        } catch (error) {
+            throw new ConusmaException("setSpeaker", "setSpeaker undefined error", error);
+        }
     }
 }
